Scene passed into ShowGuiWindow function
Previously scene was set in constructor and couldn't be changed. By passing scene as an arg to ShowGuiWindow scene will always be the correct one. Also fixed a small bug where you couldn't remove a transform component if a render component exists because the remove buttons had the same ID